Move getData logic to sheets subfolder.
